remove test.python._mockutils
