+++ qiskit/test/mock.py
@@ -181,21 +181,45 @@ def new_fake_qobj():
     )
 
 
-class FakeBackend:
-    """A fake backend."""
+class FakeTenerife:
+    """A fake 5 qubit backend."""
 
     def name(self):
-        """ name of fake backend"""
-        return 'qiskit_is_cool'
+        return 'fake_tenerife'
 
     def configuration(self):
-        """Return a make up configuration for a fake device."""
-        qx5_cmap = [[1, 0], [1, 2], [2, 3], [3, 4], [3, 14], [5, 4], [6, 5],
-                    [6, 7], [6, 11], [7, 10], [8, 7], [9, 8], [9, 10], [11, 10],
-                    [12, 5], [12, 11], [12, 13], [13, 4], [13, 14], [15, 0],
-                    [15, 2], [15, 14]]
+        cmap = [[1, 0], [2, 0], [2, 1], [3, 2], [3, 4], [4, 2]]
+
         return BackendConfiguration(
-            backend_name='fake',
+            backend_name='fake_tenerife',
+            backend_version='0.0.0',
+            n_qubits=5,
+            basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
+            simulator=False,
+            local=True,
+            conditional=False,
+            open_pulse=False,
+            memory=False,
+            max_shots=65536,
+            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
+            coupling_map=cmap,
+        )
+
+
+class FakeRueschlikon:
+    """A fake 16 qubit backend."""
+
+    def name(self):
+        return 'fake_rueschlikon'
+
+    def configuration(self):
+        camp = [[1, 0], [1, 2], [2, 3], [3, 4], [3, 14], [5, 4], [6, 5],
+                [6, 7], [6, 11], [7, 10], [8, 7], [9, 8], [9, 10],
+                [11, 10], [12, 5], [12, 11], [12, 13], [13, 4],
+                [13, 14], [15, 0], [15, 2], [15, 14]]
+
+        return BackendConfiguration(
+            backend_name='fake_rueschlikon',
             backend_version='0.0.0',
             n_qubits=16,
             basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
@@ -206,5 +230,5 @@ class FakeBackend:
             memory=False,
             max_shots=65536,
             gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
-            coupling_map=qx5_cmap,
+            coupling_map=cmap,
         )
